refactor: remove some redundant MultipleSelectorTextField props
