+++ example/lib/selectors/custom_selector.dart
@@ -57,12 +57,8 @@ class _CustomSelectorState extends State<CustomSelector> {
           initialOptions: options.getRange(0, 5).toList(),
           isMultiple: true,
           maxMenuHeight: 320,
-          fieldBuilder: (controller, onSearch, onTap) {
-            return CustomField(
-              controller: controller,
-              onTap: onTap,
-              onSearch: onSearch,
-            );
+          fieldBuilder: (_, controller) {
+            return CustomField(controller: controller);
           },
         ),
       ],
@@ -132,17 +128,14 @@ class PersonSelectorTile extends StatelessWidget {
 class CustomField extends StatelessWidget {
   const CustomField({
     Key? key,
-    required this.onTap,
     required this.controller,
-    this.onSearch,
   }) : super(key: key);
 
-  final VoidCallback onTap;
   final AdaptiveSelectorController<Person> controller;
-  final ValueChanged<String>? onSearch;
 
   @override
   Widget build(BuildContext context) {
+    final selector = AdaptiveSelector.of(context);
     return ValueListenableBuilder(
       valueListenable: controller.selectedOptionsNotifier,
       builder: (context, value, _) {
@@ -186,9 +179,9 @@ class CustomField extends StatelessWidget {
                   SizedBox(
                     width: 100,
                     child: TextField(
-                      onChanged: onSearch,
-                      onTap: onTap,
-                      readOnly: onSearch == null,
+                      onChanged: selector.handleTextChange,
+                      onTap: selector.showSelector,
+                      readOnly: selector.widget.onSearch == null,
                       decoration: const InputDecoration(
                         border: InputBorder.none,
                         focusedBorder: InputBorder.none,
